Listagem Tarefa incompleto
É necessário fazer o tratamento dos dados passados pelo array, a estratégia por enquanto é usar os foreachs e conseguir organizá-los de alguma forma. Só mexer com o index

+++ views/index.php
@@ -8,40 +8,40 @@
 <button class='butao nova-tarefa' type='button' onclick="window.location.href = '<?php echo APP.'tarefa/novo';?>'">
     Nova Tarefa
 </button>
-
+<div class='container-tarefas'>
 <?php
+    $chaves = array_keys($tarefas);
+    print_r($tarefas);
+    $cont = 0;
     foreach($tarefas as $tarefa){
-        $caminho = APP;
-        $data = implode('/', array_reverse(explode('-', $tarefa['prazo'])));
-        echo "
-            <div class='container-tarefas'>
+        echo "AAAAAAAAAAAAAA";
+        print_r($tarefa);
+        if ($tarefa == false){
+            $cont++;
+            continue;
+        }else{
+            $chave = $chaves[$cont];
+            $caminho = APP;
+            $data = implode('/', array_reverse(explode('-', $tarefa['prazo'])));
+            echo "
+                <div class='container-categoria'>
+                    <div class='categoria-nome'>
+                        $chave 
+                    </div>
+                    <hr>
+                </div>
                 <div class='tarefa'>
                     <div>
                         <p class='nome-tarefa'>{$tarefa['descricao']}</p>
-                        <p class='prazo-tarefa'>#$data</p>
+                        <p class='prazo-tarefa'>$data</p>
                     </div>
                     <a class='butao deletar-tarefa' type='button' href='$caminho/tarefa/excluir/{$tarefa['id']}'>
                         <p class='check-icon'>✓</p>
                     </a>
                 </div>
-            </div>
-        ";
+            ";
+            $cont++;
+        }
     }
 ?>
-<div class='container-categoria'>
-    <div class='categoria-nome'>
-        Tarefas de casa
-    </div>
-    <hr>
-</div>
-<div class='container-tarefas'>
-    <div class='tarefa'>
-        <div>
-            <p class='nome-tarefa'>Varrer a casa</p>
-            <p class='prazo-tarefa'>#17/02/2023</p>
-        </div>
-        <a class='butao deletar-tarefa' type='button' href='$caminho/tarefa/excluir/{$tarefa['id']}'>
-            <p class='check-icon'>✓</p>
-        </a>
-    </div>
 </div>
